html/editing/editing-0/writing-suggestions/writingsuggestions.html

<!DOCTYPE html>
<title>Tests for the writingsuggestions attribute</title>
<link rel='author' title='Sanket Joshi' href='mailto:sajos@microsoft.com'>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#writingsuggestions">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

customElements.define('test-custom-element', class extends HTMLElement {});

test(function() {
    assert_true('writingsuggestions' in document.createElement('input'));
}, 'Test that the writingsuggestions attribute is available on HTMLInputElement.');

test(function() {
    assert_true('writingsuggestions' in document.createElement('textarea'));
}, 'Test that the writingsuggestions attribute is available on HTMLTextAreaElement.');

test(function() {
    assert_true('writingsuggestions' in document.createElement('div'));
}, 'Test that the writingsuggestions attribute is available on HTMLDivElement.');

test(function() {
    assert_true('writingsuggestions' in document.createElement('span'));
}, 'Test that the writingsuggestions attribute is available on HTMLSpanElement.');

test(function() {
    assert_true('writingsuggestions' in document.createElement('test-custom-element'));
}, 'Test that the writingsuggestions attribute is available on custom elements.');

test(function() {
  const elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('span'),
                   document.createElement('test-custom-element') ];

  elements.forEach(function(element) {
    element.writingsuggestions = 'on';
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), 'on');
  });
}, 'Test setting the `writingsuggestions` attribute to `on` directly on the target element.');

test(function() {
  const elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('span'),
                   document.createElement('test-custom-element') ];

  elements.forEach(function(element) {
    element.writingsuggestions = '';
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), '');
  });
}, 'Test setting the `writingsuggestions` attribute to the empty string directly on the target element.');

test(function() {
  const elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('span'),
                   document.createElement('test-custom-element') ];

  elements.forEach(function(element) {
    element.writingsuggestions = 'foo';
    // `element.writingsuggestions` should be set to the default value, which may vary across browsers.
    assert_equals(element.getAttribute('writingsuggestions'), 'off');
  });
}, 'Test setting the `writingsuggestions` attribute to an invalid value directly on the target element.');

test(function() {
  const elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('span'),
                   document.createElement('test-custom-element') ];

  elements.forEach(function(element) {
    // `element.writingsuggestions` should be set to the default value, which may vary across browsers.
    assert_equals(element.getAttribute('writingsuggestions'), null);
  });
}, 'Test the writing suggestions state when the `writingsuggestions` attribute is missing.');

test(function() {
  const elements = [ new DOMParser().parseFromString('<input writingsuggestions />', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<textarea writingsuggestions></textarea>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<div writingsuggestions></div>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<span writingsuggestions></span>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    // `element.writingsuggestions` should be set to the default value, which may vary across browsers.
    assert_equals(element.getAttribute('writingsuggestions'), '');
  });
}, 'Test setting the `writingsuggestions` attribute with a missing value directly on the target element.');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions="on"><input /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><textarea></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><div></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><span></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'on');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), 'on');
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), null);
  });
}, 'Test setting the `writingsuggestions` attribute to `on` on a parent element.');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions=""><input /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><textarea></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><div></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><span></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'on');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), '');
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), null);
  });
}, 'Test setting the `writingsuggestions` attribute to an empty string on a parent element.');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions="off"><input /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><textarea></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><div></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><span></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'off');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), 'off');
    assert_equals(element.writingsuggestions, 'off');
    assert_equals(element.getAttribute('writingsuggestions'), null);
  });
}, 'Test setting the `writingsuggestions` attribute to `off` on a parent element.');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions="on"><input writingsuggestions="off" /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><textarea writingsuggestions="off"></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><div writingsuggestions="off"></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="on"><span writingsuggestions="off"></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'on');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), 'on');
    assert_equals(element.writingsuggestions, 'off');
    assert_equals(element.getAttribute('writingsuggestions'), 'off');
  });
}, 'Test overriding the parent element\'s `writingsuggestions` attribute from "on" to "off".');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions=""><input writingsuggestions="off" /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><textarea writingsuggestions="off"></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><div writingsuggestions="off"></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions=""><span writingsuggestions="off"></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'on');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), '');
    assert_equals(element.writingsuggestions, 'off');
    assert_equals(element.getAttribute('writingsuggestions'), 'off');
  });
}, 'Test overriding the parent element\'s `writingsuggestions` attribute from the empty string to "off".');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions="off"><input writingsuggestions="on" /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><textarea writingsuggestions="on"></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><div writingsuggestions="on"></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><span writingsuggestions="on"></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'off');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), 'off');
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), 'on');
  });
}, 'Test overriding the parent element\'s `writingsuggestions` attribute from "off" to "on".');

test(function() {
  const elements = [ new DOMParser().parseFromString('<html><body writingsuggestions="off"><input writingsuggestions="" /></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><textarea writingsuggestions=""></textarea></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><div writingsuggestions=""></div></body></html>', 'text/html').body.firstElementChild,
                     new DOMParser().parseFromString('<html><body writingsuggestions="off"><span writingsuggestions=""></span></body></html>', 'text/html').body.firstElementChild ];

  elements.forEach(function(element) {
    assert_equals(element.parentElement.writingsuggestions, 'off');
    assert_equals(element.parentElement.getAttribute('writingsuggestions'), 'off');
    assert_equals(element.writingsuggestions, 'on');
    assert_equals(element.getAttribute('writingsuggestions'), '');
  });
}, 'Test overriding the parent element\'s `writingsuggestions` attribute from "off" to the empty string.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input /><textarea></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test turning off writing suggestions for an entire document.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input writingsuggestions="on" /><textarea></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on an input element from "off" to "on".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input /><textarea writingsuggestions="on"></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a textarea element from "off" to "on".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input /><textarea></textarea><div writingsuggestions="on"></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a div element from "off" to "on".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input /><textarea></textarea><div></div><span writingsuggestions="on"></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a span element from "off" to "on".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input writingsuggestions=""><textarea></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on an input element from "off" to the empty string.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input><textarea writingsuggestions=""></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a textarea element from "off" to the empty string.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input><textarea></textarea><div writingsuggestions=""></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a div element from "off" to the empty string.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="off"><body><input><textarea></textarea><div></div><span writingsuggestions=""></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'off');
  assert_equals(doc.body.writingsuggestions, 'off');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a span element from "off" to the empty string.');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="on"><body><input writingsuggestions="off"><textarea></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on an input element from "on" to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="on"><body><input><textarea writingsuggestions="off"></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a textarea element from "on" to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="on"><body><input><textarea></textarea><div writingsuggestions="off"></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a div element from "on" to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="on"><body><input><textarea></textarea><div></div><span writingsuggestions="off"></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a span element from "on" to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions=""><body><input writingsuggestions="off"><textarea></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'off');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on an input element from the empty string to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions=""><body><input><textarea writingsuggestions="off"></textarea><div></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'off');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a textarea element from the empty string to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions=""><body><input><textarea></textarea><div writingsuggestions="off"></div><span></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'off');
  assert_equals(doc.querySelector('span').writingsuggestions, 'on');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a div element from the empty string to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions=""><body><input><textarea></textarea><div></div><span writingsuggestions="off"></span></body></html>', 'text/html');
  assert_equals(doc.documentElement.writingsuggestions, 'on');
  assert_equals(doc.body.writingsuggestions, 'on');
  assert_equals(doc.querySelector('input').writingsuggestions, 'on');
  assert_equals(doc.querySelector('textarea').writingsuggestions, 'on');
  assert_equals(doc.querySelector('div').writingsuggestions, 'on');
  assert_equals(doc.querySelector('span').writingsuggestions, 'off');
}, 'Test overriding a non-parent ancestor element\'s `writingsuggestions` attribute on a span element from the empty string to "off".');

test(function() {
  const doc = new DOMParser().parseFromString('<html writingsuggestions="on"><body><div contenteditable="true"><span>Writing suggestions allowed.</span> <span writingsuggestions="off">Writing suggestions not allowed.</span></div></body></html>', 'text/html');
  const div = doc.querySelector('div');
  const span1 = doc.querySelector('span');
  const span2 = doc.querySelector('span:last-child');
  assert_equals(div.writingsuggestions, 'on');
  assert_equals(span1.writingsuggestions, 'on');
  assert_equals(span2.writingsuggestions, 'off');
}, 'Test that for continuous text on the screen, writing suggestions may be allowed in one part but not another.');

</script>
</body>
</html>
